// apps/desktop/src/main/audio-host/wire/generated/PluginAnalysisJobStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PluginAnalysisFailure } from "./PluginAnalysisFailure";
import type { PluginAnalysisReport } from "./PluginAnalysisReport";

export type PluginAnalysisJobStatus = { "state": "running", phase: string, progress: number, } | { "state": "completed", report: PluginAnalysisReport, } | { "state": "failed", failure: PluginAnalysisFailure, };